refactor(admin): modularize admin panel
