+++ app/src/main/java/com/sukuna/animestudio/presentation/admin/AdminPanelCard.kt
@@ -1,0 +1,81 @@
+package com.sukuna.animestudio.presentation.admin
+
+import androidx.compose.foundation.clickable
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.aspectRatio
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.size
+import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.material3.Card
+import androidx.compose.material3.CardDefaults
+import androidx.compose.material3.Icon
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.painter.Painter
+import androidx.compose.ui.graphics.vector.ImageVector
+import androidx.compose.ui.unit.dp
+
+/**
+ * Simple card used in [AdminPanelScreen] to navigate to different admin features.
+ */
+@Composable
+fun AdminPanelCard(
+    title: String,
+    icon: Any,
+    description: String,
+    onClick: () -> Unit
+) {
+    Card(
+        modifier = Modifier
+            .fillMaxSize()
+            .padding(vertical = 4.dp)
+            .clickable { onClick() }
+            .aspectRatio(1f),
+        shape = RoundedCornerShape(12.dp),
+        colors = CardDefaults.cardColors(containerColor = MaterialTheme.colorScheme.surfaceVariant),
+    ) {
+        Column(
+            modifier = Modifier
+                .padding(16.dp)
+                .fillMaxSize(),
+            horizontalAlignment = Alignment.CenterHorizontally,
+            verticalArrangement = Arrangement.Center
+        ) {
+            when (icon) {
+                is Painter -> Icon(
+                    painter = icon,
+                    contentDescription = title,
+                    tint = MaterialTheme.colorScheme.onPrimary.copy(alpha = 0.5f),
+                    modifier = Modifier.size(34.dp)
+                )
+
+                is ImageVector -> Icon(
+                    imageVector = icon,
+                    contentDescription = title,
+                    tint = MaterialTheme.colorScheme.onPrimary.copy(alpha = 0.5f),
+                    modifier = Modifier.size(34.dp)
+                )
+
+                else -> error("Icon must be a Painter or ImageVector")
+            }
+            Spacer(modifier = Modifier.size(8.dp))
+            Text(
+                title,
+                style = MaterialTheme.typography.titleSmall,
+            )
+            Spacer(modifier = Modifier.size(4.dp))
+            Text(
+                description,
+                style = MaterialTheme.typography.bodyMedium,
+                modifier = Modifier.fillMaxSize()
+            )
+        }
+    }
+}
+
