+++ app/src/main/java/com/sukuna/animestudio/presentation/admin/UserManagementDialog.kt
@@ -1,0 +1,245 @@
+package com.sukuna.animestudio.presentation.admin
+
+import androidx.compose.foundation.BorderStroke
+import androidx.compose.foundation.Image
+import androidx.compose.foundation.background
+import androidx.compose.foundation.clickable
+import androidx.compose.foundation.gestures.detectTapGestures
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.size
+import androidx.compose.foundation.lazy.LazyColumn
+import androidx.compose.foundation.lazy.items
+import androidx.compose.foundation.shape.CircleShape
+import androidx.compose.material3.AlertDialog
+import androidx.compose.material3.Button
+import androidx.compose.material3.CircularProgressIndicator
+import androidx.compose.material3.DropdownMenu
+import androidx.compose.material3.DropdownMenuItem
+import androidx.compose.material3.ExperimentalMaterial3Api
+import androidx.compose.material3.Icon
+import androidx.compose.material3.OutlinedButton
+import androidx.compose.material3.OutlinedTextField
+import androidx.compose.material3.Snackbar
+import androidx.compose.material3.Surface
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.LaunchedEffect
+import androidx.compose.runtime.collectAsState
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.remember
+import androidx.compose.runtime.saveable.rememberSaveable
+import androidx.compose.runtime.setValue
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.input.pointer.pointerInput
+import androidx.compose.ui.layout.ContentScale
+import androidx.compose.ui.text.style.TextOverflow
+import androidx.compose.ui.unit.dp
+import androidx.compose.ui.window.Dialog
+import androidx.compose.ui.window.DialogProperties
+import androidx.hilt.navigation.compose.hiltViewModel
+import coil.compose.rememberAsyncImagePainter
+import com.sukuna.animestudio.R
+import com.sukuna.animestudio.domain.model.User
+import com.sukuna.animestudio.domain.model.UserRole
+
+/**
+ * Dialog for managing users. Includes role update and delete actions.
+ */
+@OptIn(ExperimentalMaterial3Api::class)
+@Composable
+fun UserManagementDialog(
+    onDismiss: () -> Unit,
+    viewModel: AdminPanelViewModel = hiltViewModel()
+) {
+    val users by viewModel.users.collectAsState()
+    val isLoading by viewModel.isLoading.collectAsState()
+    val error by viewModel.error.collectAsState()
+    var search by rememberSaveable { mutableStateOf("") }
+    var showDeleteDialog by rememberSaveable { mutableStateOf<User?>(null) }
+    var snackbarMessage by remember { mutableStateOf<String?>(null) }
+
+    LaunchedEffect(Unit) { viewModel.fetchUsers() }
+
+    Dialog(
+        onDismissRequest = onDismiss,
+        properties = DialogProperties(usePlatformDefaultWidth = false)
+    ) {
+        Surface(
+            shape = MaterialTheme.shapes.large,
+            tonalElevation = 8.dp,
+            modifier = Modifier
+                .fillMaxWidth(0.95f)
+                .fillMaxWidth(0.95f)
+        ) {
+            Column(
+                modifier = Modifier.background(MaterialTheme.colorScheme.background)
+            ) {
+                Row(
+                    modifier = Modifier
+                        .fillMaxWidth()
+                        .background(MaterialTheme.colorScheme.surfaceVariant)
+                        .padding(16.dp),
+                    verticalAlignment = Alignment.CenterVertically
+                ) {
+                    Text(
+                        text = "User Management",
+                        style = MaterialTheme.typography.titleLarge,
+                        modifier = Modifier.weight(1f)
+                    )
+                    Button(onClick = onDismiss) { Text("Close") }
+                }
+                OutlinedTextField(
+                    value = search,
+                    onValueChange = { search = it },
+                    label = { Text("Search users...") },
+                    modifier = Modifier
+                        .fillMaxWidth()
+                        .padding(16.dp)
+                )
+                when {
+                    isLoading -> {
+                        Row(Modifier.fillMaxWidth(), horizontalArrangement = Arrangement.Center) {
+                            CircularProgressIndicator()
+                        }
+                    }
+                    error != null -> {
+                        Text(
+                            error!!,
+                            color = MaterialTheme.colorScheme.error,
+                            modifier = Modifier.padding(16.dp)
+                        )
+                    }
+                    else -> {
+                        val filtered = users.filter {
+                            it.username.contains(search, ignoreCase = true) ||
+                                it.email.contains(search, ignoreCase = true)
+                        }
+                        if (filtered.isEmpty()) {
+                            Text(
+                                text = "No users found.",
+                                color = MaterialTheme.colorScheme.onSurfaceVariant,
+                                modifier = Modifier.padding(16.dp)
+                            )
+                        } else {
+                            LazyColumn(
+                                verticalArrangement = Arrangement.spacedBy(8.dp),
+                                modifier = Modifier.padding(horizontal = 16.dp)
+                            ) {
+                                items(filtered, key = { it.id }) { user ->
+                                    UserManagementRow(
+                                        user = user,
+                                        onRoleChange = { newRole ->
+                                            viewModel.updateUserRole(user, newRole) { success ->
+                                                snackbarMessage = if (success) "Role updated." else "Failed to update role."
+                                            }
+                                        },
+                                        onDelete = { showDeleteDialog = user }
+                                    )
+                                }
+                            }
+                        }
+                    }
+                }
+            }
+        }
+    }
+
+    if (showDeleteDialog != null) {
+        val user = showDeleteDialog!!
+        AlertDialog(
+            onDismissRequest = { showDeleteDialog = null },
+            title = { Text("Delete User") },
+            text = { Text("Are you sure you want to delete ${user.username}?") },
+            confirmButton = {
+                Button(onClick = {
+                    viewModel.deleteUser(user) { success ->
+                        snackbarMessage = if (success) "User deleted." else "Failed to delete user."
+                        showDeleteDialog = null
+                    }
+                }) { Text("Delete") }
+            },
+            dismissButton = { Button(onClick = { showDeleteDialog = null }) { Text("Cancel") } }
+        )
+    }
+
+    if (snackbarMessage != null) {
+        Snackbar(
+            modifier = Modifier.padding(8.dp),
+            action = { Button(onClick = { snackbarMessage = null }) { Text("OK") } }
+        ) { Text(snackbarMessage!!) }
+    }
+}
+
+@Composable
+private fun UserManagementRow(
+    user: User,
+    onRoleChange: (UserRole) -> Unit,
+    onDelete: () -> Unit
+) {
+    var expanded by remember { mutableStateOf(false) }
+    Row(
+        modifier = Modifier
+            .padding(8.dp)
+            .fillMaxWidth()
+            .background(MaterialTheme.colorScheme.surfaceVariant, shape = CircleShape)
+            .padding(12.dp)
+            .pointerInput(Unit) { detectTapGestures(onLongPress = { onDelete() }) },
+        verticalAlignment = Alignment.CenterVertically
+    ) {
+        Image(
+            painter = rememberAsyncImagePainter(user.profilePictureUrl.ifEmpty { R.drawable.ic_launcher_foreground }),
+            contentDescription = "Avatar",
+            modifier = Modifier
+                .size(40.dp)
+                .background(MaterialTheme.colorScheme.primary, CircleShape),
+            contentScale = ContentScale.Crop
+        )
+        Spacer(Modifier.size(12.dp))
+        Column(Modifier.weight(1f)) {
+            Text(user.username, style = MaterialTheme.typography.titleMedium)
+            Text(
+                user.email,
+                style = MaterialTheme.typography.bodySmall,
+                color = MaterialTheme.colorScheme.onSurfaceVariant,
+                overflow = TextOverflow.Ellipsis,
+                maxLines = 1,
+                modifier = Modifier.fillMaxWidth(0.8f)
+            )
+        }
+        Box {
+            OutlinedButton(
+                onClick = { expanded = true },
+                modifier = Modifier.width(120.dp),
+                border = BorderStroke(2.dp, MaterialTheme.colorScheme.tertiary.copy(alpha = 0.8f)),
+            ) {
+                Text(
+                    user.role.name,
+                    style = MaterialTheme.typography.bodyMedium,
+                    maxLines = 1,
+                    overflow = TextOverflow.Ellipsis,
+                    color = MaterialTheme.colorScheme.tertiary.copy(alpha = 0.8f)
+                )
+            }
+            DropdownMenu(expanded = expanded, onDismissRequest = { expanded = false }) {
+                UserRole.entries.forEach { role ->
+                    DropdownMenuItem(
+                        text = { Text(role.name) },
+                        onClick = {
+                            expanded = false
+                            if (role != user.role) onRoleChange(role)
+                        }
+                    )
+                }
+            }
+        }
+    }
+}
+
